Praktikum komunikacijskih sustava/Prvi ciklus (SDR)/requirements.txt
numpy
pandas
scikit-commpy
matplotlib
Pillow